55th Commit : Set Permission Guard Member & User

+++ app/Http/Controllers/Member/MemberAdminUserController.php
@@ -15,6 +15,15 @@ use App\Http\Requests\Member\MemberAdminUserUpdateRequest;
 
 class MemberAdminUserController extends Controller
 {
+    function __construct()
+    {
+        $this->middleware('permission:member admin user create,' . getGuardNameMember(), ['only' => ['create', 'store']]);
+        $this->middleware('permission:member admin user delete,' . getGuardNameMember(), ['only' => ['destroy']]);
+        $this->middleware('permission:member admin user index,' . getGuardNameMember(), ['only' => ['index', 'show', 'data']]);
+        $this->middleware('permission:member admin user restore,' . getGuardNameMember(), ['only' => ['restore']]);
+        $this->middleware('permission:member admin user update,' . getGuardNameMember(), ['only' => ['edit', 'update']]);
+    }
+
     /**
      * Display a listing of the resource.
      */
@@ -173,20 +182,29 @@ class MemberAdminUserController extends Controller
                     return '<div class="badge badge-danger">'  . __('admin.No Action') . '</div>';
                 } else {
                     if ($query->status == 1) {
-                        return '
-                            <a href="' . route('member.admin.edit', $query->id) . '" class="btn btn-primary btn-sm">
-                                <i class="fas fa-edit"></i>
-                            </a>
-                            <a href="' . route('member.admin.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
-                                <i class="fas fa-trash-alt"></i>
-                            </a>
-                        ';
+                        if (canAccess(['member admin user update'])) {
+                            $update = '
+                                <a href="' . route('member.admin.edit', $query->id) . '" class="btn btn-primary btn-sm">
+                                    <i class="fas fa-edit"></i>
+                                </a>
+                            ';
+                        }
+                        if (canAccess(['member admin user delete'])) {
+                            $delete = '
+                                <a href="' . route('member.admin.destroy', $query->id) . '" class="btn btn-danger btn-sm delete_item">
+                                    <i class="fas fa-trash-alt"></i>
+                                </a>
+                            ';
+                        }
+                        return (!empty($update) ? $update : '') . (!empty($delete) ? $delete : '');
                     } else {
-                        return '
-                            <a href="' . route('member.admin.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('admin.Restore to Active') . '">
-                                <i class="fas fa-undo"></i>
-                            </a>
-                        ';
+                        if (canAccess(['member admin user restore'])) {
+                            return '
+                                <a href="' . route('member.admin.restore', $query->id) . '" class="btn btn-warning btn-sm" data-toggle="tooltip" title="' . __('Restore to Active') . '">
+                                    <i class="fas fa-undo"></i>
+                                </a>
+                            ';
+                        }
                     }
                 }
             })
